add SessionHelper
